requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-generativeai>=0.8.0
pydantic>=2.0.0
websockets>=12.0
pybullet>=3.2.6
numpy>=1.24.0
